progress

+++ download_winds.py
@@ -15,36 +15,41 @@ logging.basicConfig(level=logging.DEBUG, format='%(asctime)s :: %(levelname)s ::
 
 #product for the 218 12k grid with 3hourly winds, taken from the main website here:
 #https://www.ncei.noaa.gov/products/weather-climate-models/north-american-mesoscale
-url_base = 'https://www.ncei.noaa.gov/data/north-american-mesoscale-model/access/forecast/'
+
 
 def download_latest():
+    url_base = 'https://www.ncei.noaa.gov/data/north-american-mesoscale-model/access/forecast/'
 
-    logging.debug("entering function download_latest()")
+    logging.debug("Entering function download_latest()")
     #cleanup all existing files for space savings...
 
     logging.debug("Changing directory to downloaded_data/latest ")
     os.chdir('downloaded_data/latest')
 
     files = glob.glob('*.grb2')
-    logging.debug("Found files for deletion: {files}").format(files=files)
-    for f in files:
-        try:
-            os.remove(f)
-            logging.debug("Removed file {f}").format(f=f)
-        except OSError as e:
-            print("Error: %s : %s" % (f, e.strerror))
+    if files:
+        logging.debug("Found files for deletion: {files}".format(files=files))
+        for f in files:
+            try:
+                os.remove(f)
+                logging.debug("Removed file {f}".format(f=f))
+            except OSError as e:
+                print("Error: %s : %s" % (f, e.strerror))
 
 
+    logging.debug("Fetching html for {url_base}".format(url_base=url_base))
     html_buff = pd.read_html(url_base)  #download the base page for forecast months
     monthly_forecasts= html_buff[0]
     monthly_forecasts.dropna(how='all', inplace=True)
     our_month = monthly_forecasts.iloc[-1,0]  #grabs the very last forecast month..  eg. '202208/'
     #repeat procedure for latest forecast of the month...
+    logging.debug("Fetching html for {url}".format(url=our_month))
     html_buff = pd.read_html(url_base+our_month)  #download the base page for forecast months
     daily_forecasts= html_buff[0]
     daily_forecasts.dropna(how='all', inplace=True)
     our_day = daily_forecasts.iloc[-1,0]  #grabs the very last forecast day..  eg. '20220806/'
     #repeat to find the last run hourly forecast in the list..
+    logging.debug("Fetching html for {url}".format(url=our_day))
     html_buff = pd.read_html(url_base+our_month+our_day)
     hourly_forecasts= html_buff[0]
     hourly_forecasts.dropna(how='all', inplace=True)
@@ -61,13 +66,14 @@ def download_latest():
     index = 0
     for i in range(53):
         st2 = time.time()  #our program start time...
-        logging.info("Fetching file {str} ...",format(str=url_base+our_month+our_day+file_prefix+'_'+str(index).zfill(3)+file_post)
+        our_file=str(url_base+our_month+our_day+file_prefix+'_'+str(index).zfill(3)+file_post)
+        logging.info("Fetching file {file} ...".format(file=our_file))
         urllib.request.urlretrieve(url_base+our_month+our_day+file_prefix+'_'+str(index).zfill(3)+file_post, filename=file_prefix+'_'+str(i).zfill(3)+file_post)
         # get the end time
         et = time.time()
         # get the execution time
         elapsed_time = et - st2
-        print('Download took: ', elapsed_time, ' seconds')
+        logging.info('Download took: {elapsed_time} seconds'.format(elapsed_time=elapsed_time))
         if index < 36:
             index+=1
         else:
@@ -76,6 +82,7 @@ def download_latest():
 
     ft = time.time()
     elapsed_time = et - st
-    print('Final download time took: ', elapsed_time, ' seconds')
+    logging.info('Total download took: {elapsed_time} seconds'.format(elapsed_time=elapsed_time))
 
+    logging.debug("Exiting function download_latest() successfully")
 download_latest()
